made viewmodels for artists and clients also made dropdown for them

+++ StudioMetrics/Controllers/ArtistsController.cs
@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using StudioMetrics.Data;
 using StudioMetrics.Models;
+using StudioMetrics.Models.ViewModels;
 
 namespace StudioMetrics.Controllers
 {
@@ -59,9 +60,28 @@ namespace StudioMetrics.Controllers
 
         // GET: Artists/Create
         [Authorize]
-        public IActionResult Create()
+        public async Task<IActionResult> Create()
         {
-            return View();
+            var user = await GetCurrentUserAsync();
+
+            var clients = await _context.Client.Where(c => c.User == user).ToListAsync();
+
+            var clientListOptions = new List<SelectListItem>();
+
+            foreach (Client c in clients)
+            {
+                clientListOptions.Add(new SelectListItem
+                {
+                    Value = c.ClientId.ToString(),
+                    Text = c.Name
+                });
+            }
+
+            ArtistCreateViewModel createViewModel = new ArtistCreateViewModel();
+
+            createViewModel.AvailableClients = clientListOptions;
+
+            return View(createViewModel);
         }
 
         // POST: Artists/Create
@@ -70,24 +90,36 @@ namespace StudioMetrics.Controllers
         [HttpPost]
         [Authorize]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ArtistId,Name,UserId")] Artist artist)
+        public async Task<IActionResult> Create(ArtistCreateViewModel createArtist)
         {
-            ModelState.Remove("User");
-            ModelState.Remove("UserId");
+            ModelState.Remove("Artist.User");
+            ModelState.Remove("Artist.UserId");
             if (ModelState.IsValid)
             {
-                artist.User = await GetCurrentUserAsync();
-                artist.UserId = artist.User.Id;
-                _context.Add(artist);
+                createArtist.Artist.User = await GetCurrentUserAsync();
+                createArtist.Artist.UserId = createArtist.Artist.User.Id;
+                _context.Add(createArtist.Artist);
+                if (createArtist.SelectedClients != null)
+                {
+                    foreach (int clientId in createArtist.SelectedClients)
+                    {
+                        ClientArtist newCA = new ClientArtist()
+                        {
+                            ClientId = clientId,
+                            ArtistId = createArtist.Artist.ArtistId
+                        };
+                        _context.Add(newCA);
+                    }
+                }
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            return View(artist);
+            return View(createArtist);
         }
 
         // GET: Artists/Edit/5
         [Authorize]
-        public async Task<IActionResult> Edit(int? id)
+        public async Task<IActionResult> Edit(int? id, ArtistEditViewModel editViewModel)
         {
             if (id == null)
             {
@@ -99,7 +131,33 @@ namespace StudioMetrics.Controllers
             {
                 return NotFound();
             }
-            return View(artist);
+            var user = await GetCurrentUserAsync();
+
+            var clients = await _context.Client.Where(c => c.User == user).ToListAsync();
+            var clientArtists = await _context.ClientArtist.Where(ca => ca.ArtistId == id).ToListAsync();
+
+            var clientListOptions = new List<SelectListItem>();
+            var clientIds = new List<int>();
+
+            foreach (Client c in clients)
+            {
+                clientListOptions.Add(new SelectListItem
+                {
+                    Value = c.ClientId.ToString(),
+                    Text = c.Name
+                });
+            }
+
+            foreach (ClientArtist ca in clientArtists)
+            {
+                clientIds.Add(ca.ClientId);
+            }
+
+            editViewModel.AvailableClients = clientListOptions;
+            editViewModel.Artist = artist;
+            editViewModel.SelectedClients = clientIds;
+
+            return View(editViewModel);
         }
 
         // POST: Artists/Edit/5
@@ -108,27 +166,49 @@ namespace StudioMetrics.Controllers
         [HttpPost]
         [Authorize]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ArtistId,Name,UserId")] Artist artist)
+        public async Task<IActionResult> Edit(int id, ArtistEditViewModel editArtist)
         {
-            if (id != artist.ArtistId)
+            if (id != editArtist.Artist.ArtistId)
             {
                 return NotFound();
             }
 
-            ModelState.Remove("User");
-            ModelState.Remove("UserId");
+            ModelState.Remove("Artist.User");
+            ModelState.Remove("Artist.UserId");
             if (ModelState.IsValid)
             {
+                // Delete joiner tables
+                var clientArtists = await _context.ClientArtist.Where(ca => ca.ArtistId == id).ToListAsync();
+                if (clientArtists != null)
+                {
+                    foreach (ClientArtist ca in clientArtists)
+                    {
+                        _context.ClientArtist.Remove(ca);
+                    }
+                }
                 try
                 {
-                    artist.User = await GetCurrentUserAsync();
-                    artist.UserId = artist.User.Id;
-                    _context.Update(artist);
+                    editArtist.Artist.User = await GetCurrentUserAsync();
+                    editArtist.Artist.UserId = editArtist.Artist.User.Id;
+                    // Add the new joiner tables if their are any
+                    if (editArtist.SelectedClients != null)
+                    {
+                        foreach (int clientId in editArtist.SelectedClients)
+                        {
+                            ClientArtist newCA = new ClientArtist()
+                            {
+                                ClientId = clientId,
+                                ArtistId = editArtist.Artist.ArtistId
+                            };
+                            _context.Add(newCA);
+                        }
+                    }
+                    _context.Update(editArtist.Artist);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!ArtistExists(artist.ArtistId))
+                    if (!ArtistExists(editArtist.Artist.ArtistId))
                     {
                         return NotFound();
                     }
@@ -139,7 +219,7 @@ namespace StudioMetrics.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
-            return View(artist);
+            return View(editArtist);
         }
 
         // GET: Artists/Delete/5
@@ -169,13 +249,23 @@ namespace StudioMetrics.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var artist = await _context.Artist
-                .Include(p => p.ArtistProjects)
                 .SingleOrDefaultAsync(p => p.ArtistId == id);
-            //.FindAsync(id);
 
-            foreach (ArtistProject ap in artist.ArtistProjects)
+            var artistProjects = await _context.ArtistProject.Where(ap => ap.ArtistId == id).ToListAsync();
+            if (artistProjects != null)
             {
-                _context.ArtistProject.Remove(ap);
+                foreach (ArtistProject ap in artistProjects)
+                {
+                    _context.ArtistProject.Remove(ap);
+                }
+            }
+            var clientArtists = await _context.ClientArtist.Where(ap => ap.ArtistId == id).ToListAsync();
+            if (clientArtists != null)
+            {
+                foreach (ClientArtist ca in clientArtists)
+                {
+                    _context.ClientArtist.Remove(ca);
+                }
             }
 
             _context.Artist.Remove(artist);
